refactoring the navbar components into a new wrapper component thats going to keep it all in
